added more exhaustive log1p test

+++ test/numpy_lib/test_ufunc_math.py
@@ -5,10 +5,13 @@ from parakeet.testing_helpers import run_local_tests, expect
 
 int_vec = np.array([2,3])
 int_mat = np.array([int_vec, int_vec])
+
 bool_vec = np.array([True, True])
 bool_mat = np.array([bool_vec, bool_vec])
-float32_vec = np.array([1.0, 2.0])
+
+float32_vec = np.array([1.0, 2.0], dtype='float32')
 float32_mat = np.array([float32_vec, float32_vec])
+
 float64_vec = np.array([-1.0, 100000.0])
 float64_mat = np.array([float64_vec, float64_vec])
